no delete distfiles&packages on tester

+++ 700_clear_system.sh
@@ -11,12 +11,12 @@ ${sudo_cmd} cp  ${verbose} logfile ${new_root}/root
 #${sudo_cmd} cp  ${verbose} first* ${new_root}/root
 
 [ -f ${new_root}/.config ] && ${sudo_cmd} rm  ${verbose} ${new_root}/.config
-if [ $use_packages = 1 ] ; then
+if [ $rm_packages = 1 ] ; then
     ${sudo_cmd} rm  ${verbose} ${new_root}/etc/portage/make.conf/binpkg
+    ${sudo_cmd} rm  ${verbose} -Rf ${new_root}/var/cache/binpkgs/*
 fi
 
-${sudo_cmd} rm  ${verbose} -Rf ${new_root}/var/cache/distfiles/*
-${sudo_cmd} rm  ${verbose} -Rf ${new_root}/var/cache/binpkgs/*
+[ $rm_distfiles = 1 ] && ${sudo_cmd} rm  ${verbose} -Rf ${new_root}/var/cache/distfiles/*
 [ $rm_portages = 1 ] && ${sudo_cmd} rm  ${verbose} -Rf ${new_root}/var/db/repos/gentoo/*
